Refactor and improve the Period#chunks method to handle partials better.
Rather than raise an error when a period is too small to make up any chunks
and partials are not allowed, just return an empty list.  Use the chunk
methods from fat_core to generalize the algorithm.  Handle partial_first and
partial_last properly.  Was raising an error for small periods with partials
allowed.